refactored delete clothing item && added cors

+++ controllers/clothingItems.js
@@ -3,6 +3,7 @@ const {
   invalidDataPassError,
   notExistingError,
   defaultError,
+  forbiddenError,
 } = require("../utils/errors");
 
 function getClothingItems(req, res) {
@@ -18,21 +19,31 @@ function getClothingItems(req, res) {
     );
 }
 function deleteClothingItem(req, res) {
-  clothingItemModel
-    .findOneAndRemove({ _id: req.params.itemId })
-    .orFail()
-    .then((user) => res.status(200).send(user))
-    .catch((err) => {
-      if (err.name === "ValidationError" || err.name === "CastError") {
-        return res.status(invalidDataPassError).send({ message: err.message });
-      }
-      if (err.name === "DocumentNotFoundError") {
-        return res.status(notExistingError).send({ message: err.message });
-      }
-      return res
-        .status(defaultError)
-        .send({ message: "An error has occurred on the server" });
-    });
+  const { owner } = req.body;
+  if (owner === req._id) {
+    clothingItemModel
+      .findOneAndRemove({ _id: req.params.itemId })
+      .orFail()
+      .then((user) => res.status(200).send(user))
+      .catch((err) => {
+        if (err.name === "ValidationError" || err.name === "CastError") {
+          return res
+            .status(invalidDataPassError)
+            .send({ message: err.message });
+        }
+        if (
+          err.name === "DocumentNotFoundError" ||
+          err instanceof NotFoundError
+        ) {
+          return res.status(notExistingError).send({ message: err.message });
+        }
+        return res
+          .status(defaultError)
+          .send({ message: "An error has occurred on the server" });
+      });
+  } else {
+    return res.status(forbiddenError).send(message);
+  }
 }
 function likeItem(req, res) {
   clothingItemModel
